Server update
Build 41.61 (CEO 2)
Disabled "FireBook", forgot to commit and push to repository.

Build 41.61 (CEO 2-2)
New server save instance, direct copy of Build 41.61 (CEO 2).
The only difference was this save will properly load in the map mods, previous save did not load it in.
Map load priority is wrong, and it was a mistake on my end.